backend/requirements.txt
# Premnishani Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
boto3==1.34.14
razorpay==1.4.0
stripe==7.10.0
chromadb==0.4.22
slowapi==0.1.9
